Explicitly check for the value of 'staffedit-tag' in the request object instead of merely checking for the presence of that array item
Some things, like Lupo's AutoSubmit JS (which is used by the ImageAnnotator gadget), send 'staffedit-tag' => 'none' back to the server as a part of response.
This behavior seems to differ in practise from what standard MediaWiki EditPage does, and thus image annotation edits would incorrectly show up as staff edits, when in reality the ImageAnnotator gadget UI literally does not even *have* an option to mark an edit as a staff edit.

Change-Id: I2bc6b181f5cd069b54684302f9e1ed9525c7e457